backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1 
boto3==1.34.118
google-cloud-vision==3.10.2
pydantic-settings==2.3.4 